+++ usb/usb_host_msc/src/msc_host.c
@@ -94,6 +94,7 @@ typedef struct {
     void *user_arg;
     SemaphoreHandle_t all_events_handled;
     volatile bool end_client_event_handling;
+    bool event_handling_started;
     STAILQ_HEAD(devices, msc_host_device) devices_tailq;
 } msc_driver_t;
 
@@ -321,17 +322,33 @@ static bool is_mass_storage_device(uint8_t dev_addr)
     return is_msc_device;
 }
 
+esp_err_t msc_host_handle_events(uint32_t timeout)
+{
+    MSC_RETURN_ON_FALSE(s_msc_driver != NULL, ESP_ERR_INVALID_STATE);
+
+    ESP_LOGV(TAG, "USB MSC handling");
+    s_msc_driver->event_handling_started = true;
+    esp_err_t ret = usb_host_client_handle_events(s_msc_driver->client_handle, timeout);
+    if (s_msc_driver->end_client_event_handling) {
+        xSemaphoreGive(s_msc_driver->all_events_handled);
+        return ESP_FAIL;
+    }
+    return ret;
+}
+
+/**
+ * @brief USB Client Event handler
+ *
+ * Handle all USB client events such as USB transfers and connections/disconnections
+ *
+ * @param[in] arg   Argument, not used
+ */
 static void event_handler_task(void *arg)
 {
-    while (1) {
-        usb_host_client_handle_events(s_msc_driver->client_handle, portMAX_DELAY);
-
-        if (s_msc_driver->end_client_event_handling) {
-            break;
-        }
+    ESP_LOGD(TAG, "USB HID handling start");
+    while (msc_host_handle_events(portMAX_DELAY) == ESP_OK) {
     }
-    ESP_ERROR_CHECK( usb_host_client_deregister(s_msc_driver->client_handle) );
-    xSemaphoreGive(s_msc_driver->all_events_handled);
+    ESP_LOGD(TAG, "USB HID handling stop");
     vTaskDelete(NULL);
 }
 
@@ -445,9 +462,13 @@ esp_err_t msc_host_uninstall(void)
     s_msc_driver->end_client_event_handling = true;
     MSC_EXIT_CRITICAL();
 
-    ESP_ERROR_CHECK( usb_host_client_unblock(s_msc_driver->client_handle) );
-    xSemaphoreTake(s_msc_driver->all_events_handled, portMAX_DELAY);
+    if (s_msc_driver->event_handling_started) {
+        ESP_ERROR_CHECK( usb_host_client_unblock(s_msc_driver->client_handle) );
+        // In case the event handling started, we must wait until it finishes
+        xSemaphoreTake(s_msc_driver->all_events_handled, portMAX_DELAY);
+    }
     vSemaphoreDelete(s_msc_driver->all_events_handled);
+    ESP_ERROR_CHECK( usb_host_client_deregister(s_msc_driver->client_handle) );
     free(s_msc_driver);
     s_msc_driver = NULL;
     return ESP_OK;
@@ -513,13 +534,6 @@ esp_err_t msc_host_write_sector(msc_host_device_handle_t device, size_t sector, 
     msc_device_t *dev = (msc_device_t *)device;
 
     return scsi_cmd_write10(dev, data, sector, 1, dev->disk.block_size);
-}
-
-esp_err_t msc_host_handle_events(uint32_t timeout_ms)
-{
-    MSC_RETURN_ON_FALSE(s_msc_driver != NULL, ESP_ERR_INVALID_STATE);
-
-    return usb_host_client_handle_events(s_msc_driver->client_handle, pdMS_TO_TICKS(timeout_ms));
 }
 
 static void copy_string_desc(wchar_t *dest, const usb_str_desc_t *src)