+++ usb/usb_host_msc/test/test_msc.c
@@ -122,6 +122,22 @@ static void handle_usb_events(void *args)
     vTaskDelete(NULL);
 }
 
+/**
+ * @brief MSC driver handling task
+ *
+ * This task is only used if the MSC driver was installed with no background task
+ *
+ * @param[in] args Not used
+ */
+static void msc_task(void *args)
+{
+    ESP_LOGI(TAG, "USB MSC handling start");
+    while (msc_host_handle_events(portMAX_DELAY) == ESP_OK) {
+    }
+    ESP_LOGI(TAG, "USB MSC handling stop");
+    vTaskDelete(NULL);
+}
+
 #if ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(5, 0, 0)
 static void check_file_content(const char *file_path, const char *expected)
 {
@@ -166,7 +182,7 @@ static void check_sudden_disconnect(void)
     fclose(file);
 }
 
-static void msc_setup(void)
+static void msc_test_init(void)
 {
     BaseType_t task_created;
 
@@ -190,15 +206,10 @@ static void msc_setup(void)
 
     task_created = xTaskCreatePinnedToCore(handle_usb_events, "usb_events", 2 * 2048, NULL, 2, NULL, 0);
     TEST_ASSERT(task_created);
+}
 
-    const msc_host_driver_config_t msc_config = {
-        .create_backround_task = true,
-        .callback = msc_event_cb,
-        .stack_size = 4096,
-        .task_priority = 5,
-    };
-    ESP_OK_ASSERT( msc_host_install(&msc_config) );
-
+static void msc_test_wait_and_install_device(void)
+{
     ESP_LOGI(TAG, "Waiting for USB stick to be connected");
     msc_host_event_t app_event;
     xQueueReceive(app_queue, &app_event, portMAX_DELAY);
@@ -209,16 +220,32 @@ static void msc_setup(void)
     ESP_OK_ASSERT( msc_host_vfs_register(device, "/usb", &mount_config, &vfs_handle) );
 }
 
-static void msc_teardown(void)
+static void msc_setup(void)
 {
-    vTaskDelay(10); // Wait to finish any ongoing USB operations
+    msc_test_init();
+    const msc_host_driver_config_t msc_config = {
+        .create_backround_task = true,
+        .callback = msc_event_cb,
+        .stack_size = 4096,
+        .task_priority = 5,
+    };
+    ESP_OK_ASSERT( msc_host_install(&msc_config) );
+    msc_test_wait_and_install_device();
+}
 
+static void msc_test_uninstall_device(void)
+{
     ESP_OK_ASSERT( msc_host_vfs_unregister(vfs_handle) );
     ESP_OK_ASSERT( msc_host_uninstall_device(device) );
+}
+
+static void msc_test_deinit(void)
+{
     ESP_OK_ASSERT( msc_host_uninstall() );
 
     xSemaphoreTake(ready_to_deinit_usb, portMAX_DELAY);
     vSemaphoreDelete(ready_to_deinit_usb);
+    vTaskDelay(10); // Wait to finish any ongoing USB operations
     ESP_OK_ASSERT( usb_host_uninstall() );
     //Tear down USB PHY
     ESP_OK_ASSERT(usb_del_phy(phy_hdl));
@@ -226,6 +253,12 @@ static void msc_teardown(void)
 
     vQueueDelete(app_queue);
     vTaskDelay(10); // Wait for FreeRTOS to clean up deleted tasks
+}
+
+static void msc_teardown(void)
+{
+    msc_test_uninstall_device();
+    msc_test_deinit();
 }
 
 static void write_read_sectors(void)
@@ -451,6 +484,49 @@ TEST_CASE("device_info", "[usb_msc]")
     msc_teardown();
     TEST_ASSERT_EQUAL(ESP_OK, err);
     print_device_info(&info);
+}
+
+/**
+ * @brief USB MSC driver with no background task
+ *
+ * Install the driver without background task
+ * and make sure that everything works
+ */
+TEST_CASE("no_background_task", "[usb_msc]")
+{
+    msc_test_init();
+    const msc_host_driver_config_t msc_config = {
+        .create_backround_task = false,
+        .callback = msc_event_cb,
+        .stack_size = 4096,
+        .task_priority = 5,
+    };
+    ESP_OK_ASSERT( msc_host_install(&msc_config) );
+    BaseType_t task_created = xTaskCreatePinnedToCore(msc_task, "msc_events", 2 * 2048, NULL, 2, NULL, 0);
+    TEST_ASSERT(task_created);
+    msc_test_wait_and_install_device();
+    write_read_sectors(); // Do some dummy operations
+    msc_teardown();
+}
+
+/**
+ * @brief USB MSC driver with no background task
+ *
+ * Install and uninstall the driver without background task
+ * without ever calling usb_host_client_handle_events()
+ */
+TEST_CASE("no_background_task_2", "[usb_msc]")
+{
+    msc_test_init();
+    const msc_host_driver_config_t msc_config = {
+        .create_backround_task = false,
+        .callback = msc_event_cb,
+    };
+    ESP_OK_ASSERT( msc_host_install(&msc_config) );
+
+    vTaskDelay(100); // Give USB Host Library some time for device enumeration
+
+    msc_test_deinit();
 }
 
 /**